refactor(archunit): extract shared validation helper and centralize tsgo scripts
- Extract compilerOptions validation logic into shared helper (validate-compiler-options.ts)
  to eliminate duplication between @cellix/archunit-tests and @ocom/archunit-tests
- Make JWT/OAuth2 payload interface fields required for stronger type checking
  (TestIdTokenPayload, TestClaimsPayload, TestReservedClaimsPayload)
- Centralize tsgo script wrappers in @cellix/config-typescript package.json
  for consistent flag management across workspace packages
- Fix type assertion in multi-registration.test.ts to handle Record<string, unknown> cast

All tests passing. Addresses code review feedback on TS 7.0 upgrade work.
