+++ build-pipeline/core/monorepo-build-stage.yml
@@ -54,10 +54,10 @@ stages:
       # SonarCloud documentation https://docs.sonarsource.com/sonarqube-server/latest/analyzing-source-code/scm-integration/#:~:text=A%20full%20clone%20is%20required%20for%20this%20integration%20to%20be%20able%20to%20collect%20the%20required%20blame%20information%20(see%20Known%20Issues).%20If%20a%20shallow%20clone%20is%20detected%2C%20a%20warning%20will%20be%20logged%20and%20no%20attempt%20will%20be%20made%20to%20retrieve%20blame%20information.
 
     # Ensure the correct version of Node is installed.
-    - task: NodeTool@0
+    - task: UseNode@1
       displayName: 'Install: Node.js - 22.22.2'
       inputs:
-        versionSpec: '22.22.2'
+        version: '22.22.2'
 
     # Cache Azure Functions Core Tools
     - task: Cache@2
